backend_birthday_remainder_app/requirements.txt
Django==4.0.2
gunicorn==20.0.4
djangorestframework==3.13.1
psycopg2-binary>=2.8
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.0.0
